Revert register hook
